refactor(entities): move edges and file bodies into aboveView (Phase D)
Edge bodies, anchor dots, and file entity bodies relocated to aboveView.
