--- src/renderer/canvas-bg/EntityBlockLayers.tsx
@@ -1,5 +1,0 @@
-// Facade: re-exports per-type entity layer components. Sticky text bodies
-// migrated to aboveView (`above-view/StickyBodyLayer.tsx`) in Phase C of the
-// aboveView migration; the rest still live in canvas-bg.
-export { FileBlockLayer } from './FileBlockLayer'
-export { GroupBoundsLayer } from './GroupBoundsLayer'
